# PeopleCounter_Manager/docs/swagger.yaml
basePath: /
definitions:
...
  PeopleCounter_Manager_internal_models.User:
    properties:
      description:
        type: string
      email:
        type: string
      id:
        type: integer
      login:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  PeopleCounter_Manager_internal_models.UserCountersRequest:
    properties:
      counter_ids:
        items:
          type: integer
        type: array
    type: object
  PeopleCounter_Manager_internal_models.UserCreateRequest:
    properties:
      description:
        type: string
      email:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        description: '"admin" или "viewer"'
        type: string
    type: object
  PeopleCounter_Manager_internal_models.UserUpdateRequest:
    properties:
      description:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        description: 'Указатель: если nil, пароль не меняется'
        type: string
      role:
        type: string
    type: object
host: localhost:9000
info:
  contact:
    email: support@example.com
    name: API Support
  description: REST API для управления счетчиками посетителей и сбора статистики
  termsOfService: http://swagger.io/terms/
  title: People Counter Manager API
  version: "1.0"
paths:
...
  /api/users:
    get:
      description: Возвращает список всех зарегистрированных пользователей системы
        (без хэшей паролей).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/PeopleCounter_Manager_internal_models.User'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить список пользователей
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Добавляет нового пользователя в систему (только для администраторов).
        Пароль хэшируется перед сохранением.
      parameters:
      - description: Данные нового пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PeopleCounter_Manager_internal_models.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: '{\"id\": 1}'
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Неверный формат JSON
          schema:
            type: string
        "500":
          description: Ошибка сервера (возможно логин уже занят)
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создать нового пользователя
      tags:
      - Users
  /api/users/{id}:
    delete:
      description: Удаляет пользователя из системы. Все привязки пользователя к камерам
        удаляются каскадно.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление
        "400":
          description: Неверный ID
          schema:
            type: string
        "500":
          description: Ошибка удаления
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить пользователя
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Обновляет профиль пользователя и его роль. Если передан новый пароль,
        он будет захэширован и обновлен.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PeopleCounter_Manager_internal_models.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"status\":\"success\"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID или формат JSON
          schema:
            type: string
        "500":
          description: Ошибка обновления
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить данные пользователя
      tags:
      - Users
  /api/users/{id}/counters:
    get:
      description: Возвращает массив ID камер, к которым данному пользователю предоставлен
        доступ.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"counter_ids\": [1, 2, 3]}'
          schema:
            additionalProperties:
              items:
                type: integer
              type: array
            type: object
        "400":
          description: Неверный ID
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить доступные камеры пользователя
      tags:
      - Users Access
    post:
      consumes:
      - application/json
      description: Полностью перезаписывает список камер, доступных пользователю.
        Старые привязки удаляются.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Массив ID камер
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PeopleCounter_Manager_internal_models.UserCountersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"status\":\"success\"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID или формат JSON
          schema:
            type: string
        "500":
          description: Ошибка при сохранении прав
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Назначить камеры пользователю
      tags:
      - Users Access
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"